Create 21. Merge Two Sorted Lists.cs
This solution uses recursion to merge two sorted linked lists into a single sorted linked list. The function takes two linked lists as input (list1 and list2) and returns a new linked list that contains all the elements from both input lists in sorted order.

The function first checks if either of the input lists is null. If one of them is null, it returns the other list as is.

If neither list is null, it compares the first elements of both lists (list1.val and list2.val) and chooses the smaller one as the head of the merged list. It then recursively calls itself with the remaining elements of the chosen list and the other list until one of them becomes null.